First Bugsnag#notify call from React Native JS.
I'll need to add some steps to the Readme, but it's essentially:

1) Integrate KSCrash and Bugsnag into your application Xcode project.

2) Add "react-native-bugsnag" dependency to your package.json.

3) Then do sample exception like this:

var Bugsnag = require('react-native-bugsnag');

Bugsnag.notify("ExceptionName1", "ExceptionReason1", null);

This is the start of the rest of it. I've commented out all the other code for now so we can get this working and then uncomment and modify the other methods going forward.

It's hitting midnight here so I was only able to test a single JS excpetion and it doesn't give us much context because it comes over the bridge and we lose the location of the original error.

So, we might want to use the JS library when **handling excpetions** in React Native. Will have to test a lot more on how to get the best results.

But, it's already nice to get native iOS exceptions in the beautiful and familiar Bugsnag interface.